# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agent-api",
    "version": "1.0.0",
    "description": "Open-source web scraper, crawler & screenshot API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/crawl": {
      "post": {
        "summary": "Start a crawl job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string" },
                  "maxPages": { "type": "number", "default": 10 },
                  "format": { "type": "string", "default": "markdown" }
                },
                "required": ["url"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job started"
          }
        }
      }
    },
    "/v1/crawl/{jobId}": {
      "get": {
        "summary": "Poll crawl job status",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status and results"
          }
        }
      }
    },
...
  }
}